Move CommandSearcher classes
