+++ app/src/main/java/net/nhiroki/bluelineconsole/applicationMain/PreferencesActivity.java
@@ -1,19 +1,17 @@
 package net.nhiroki.bluelineconsole.applicationMain;
 
 import android.Manifest;
-import android.app.Activity;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.view.Gravity;
 
-import androidx.fragment.app.Fragment;
 import androidx.preference.PreferenceManager;
 import androidx.preference.SwitchPreference;
 
 import net.nhiroki.bluelineconsole.R;
-import net.nhiroki.bluelineconsole.commandSearchers.ContactSearchCommandSearcher;
+import net.nhiroki.bluelineconsole.commandSearchers.eachSearcher.ContactSearchCommandSearcher;
 import net.nhiroki.bluelineconsole.wrapperForAndroid.ContactsReader;
 
 public class PreferencesActivity extends BaseWindowActivity {
