+++ app/src/main/java/net/nhiroki/bluelineconsole/commandSearchers/eachSearcher/CalculatorCommandSearcher.java
@@ -1,4 +1,4 @@
-package net.nhiroki.bluelineconsole.commandSearchers;
+package net.nhiroki.bluelineconsole.commandSearchers.eachSearcher;
 
 import android.content.Context;
 import android.graphics.Typeface;
