+++ app/src/main/java/net/nhiroki/bluelineconsole/commandSearchers/eachSearcher/SearchEngineDefaultCommandSearcher.java
@@ -1,4 +1,4 @@
-package net.nhiroki.bluelineconsole.commandSearchers;
+package net.nhiroki.bluelineconsole.commandSearchers.eachSearcher;
 
 import android.content.Context;
 import android.content.Intent;
@@ -20,10 +20,10 @@ import net.nhiroki.bluelineconsole.interfaces.EventLauncher;
 import java.util.ArrayList;
 import java.util.List;
 
-public class SearchEngineCommandSearcher implements CommandSearcher {
+public class SearchEngineDefaultCommandSearcher implements CommandSearcher {
     private final WebSearchEnginesDatabase _searchEngineDB;
 
-    public SearchEngineCommandSearcher(Context context) {
+    public SearchEngineDefaultCommandSearcher(Context context) {
         this._searchEngineDB = new WebSearchEnginesDatabase(context);
         this.refresh(context);
     }
@@ -49,85 +49,12 @@ public class SearchEngineCommandSearcher implements CommandSearcher {
     public List<CandidateEntry> searchCandidateEntries(String s, Context context) {
         List<CandidateEntry> cands = new ArrayList<>();
 
-        if (s.contains(" ")){
-            int split = s.indexOf(' ');
-            String engine = s.substring(0, split).toLowerCase();
-            String query = s.substring(split + 1);
+        WebSearchEngine searchEngine = this._searchEngineDB.getDefaultEngineByPreference(context);
 
-            if (!engine.equals("")) {
-                List<WebSearchEngine> searchEngines = this._searchEngineDB.searchEngineListByNameQuery(context, engine);
-
-                for (WebSearchEngine e: searchEngines) {
-                    cands.add(new SearchEngineCandidateEntry(context, query, e.display_name, e.url_base));
-                }
-            }
-
-        } else {
-            if (!s.equals("")) {
-                List<WebSearchEngine> urls = this._searchEngineDB.searchStaticPageListByNameQuery(context, s);
-                for (WebSearchEngine e: urls) {
-                    cands.add(new StaticPageCandidateEntry(context, e.display_name, e.url_base));
-                }
-            }
+        if (searchEngine != null) {
+            cands.add(new SearchEngineCandidateEntry(context, s, searchEngine.display_name, searchEngine.url_base));
         }
         return cands;
-    }
-
-    private static class StaticPageCandidateEntry implements CandidateEntry {
-        private final String urlBase;
-        private final String title;
-
-        StaticPageCandidateEntry(Context context, String pageName, String urlBase) {
-            this.urlBase = urlBase;
-            this.title = String.format(context.getString(R.string.result_static_webpage_format), pageName);
-        }
-
-        @Override
-        public boolean hasLongView() {
-            return false;
-        }
-
-        @Override
-        @NonNull
-        public String getTitle() {
-            return title;
-        }
-
-        @Override
-        public View getView(Context context) {
-            return null;
-        }
-
-        @Override
-        public EventLauncher getEventLauncher(Context context) {
-            return new EventLauncher() {
-                @Override
-                public void launch(BaseWindowActivity activity) {
-                    activity.startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse(urlBase)));
-                    activity.finish();
-                }
-            };
-        }
-
-        @Override
-        public Drawable getIcon(Context context) {
-            return ContextCompat.getDrawable(context, android.R.drawable.ic_menu_compass);
-        }
-
-        @Override
-        public boolean hasEvent() {
-            return true;
-        }
-
-        @Override
-        public boolean isSubItem() {
-            return false;
-        }
-
-        @Override
-        public boolean viewIsRecyclable() {
-            return true;
-        }
     }
 
     private static class SearchEngineCandidateEntry implements CandidateEntry {
